Fix a bug in the sequence during duplication.
The bug occured when an input socket of a task (t2) was bound to
an other task (t1) which was not in the tasks graph. If t1 was in
a common module of a task (t3) which is in the tasks graph, then
the binding was made between t1 and t2 when it should not.

This commit includes a check before the automatic binding: if a
task is bound to an other task (which is not included in the tasks
graph) then the binding in the duplication is skipped.